refactor: cli demo,templates -> fixtures
