--- tests/test-metadata.xml
@@ -1,38 +1,0 @@
-<record xmlns="http://www.loc.gov/MARC21/slim">
-  <controlfield tag='001'>990079940640203331</controlfield>
-  <controlfield tag='003'>AT-OBV</controlfield>
-  <controlfield tag='005'>20170703041800.0</controlfield>
-  <controlfield tag='007'>cr</controlfield>
-  <controlfield tag='008'>100504|1932</controlfield>
-  <controlfield tag='009'>AC08088803</controlfield>
-  <datafield tag='035' ind1=' ' ind2=' '>
-    <subfield code='a'>AC08088803</subfield>
-  </datafield>
-  <datafield tag='035' ind1=' ' ind2=' '>
-    <subfield code='a'>(AT-OBV)AC08088803</subfield>
-    <subfield code='a'>(Aleph)007994064ACC01</subfield>
-    <subfield code='a'>(DE-599)OBVAC08088803</subfield>
-  </datafield>
-  <datafield tag='100' ind1='1' ind2=' '>
-    <subfield code='a'>Philipp</subfield>
-  </datafield>
-  <datafield tag='245' ind1='0' ind2='0'>
-    <subfield code='a'>&lt;&lt;Die&gt;&gt; Internationale Werkbundsiedlung Wien 1932</subfield>
-    <subfield code='c'>hrsg. von Josef Frank</subfield>
-  </datafield>
-  <datafield tag='264' ind1=' ' ind2='0'>
-    <subfield code='a'>Wien</subfield>
-    <subfield code='b'>Schroll</subfield>
-    <subfield code='c'>1932</subfield>
-  </datafield>
-  <datafield tag='260' ind1='3' ind2=' '>
-    <subfield code='b'>TU Graz</subfield>
-  </datafield>
-  <datafield tag='362' ind1='0' ind2=' '>
-    <subfield code='a'>2022</subfield>
-  </datafield>
-  <datafield tag='502' ind1=' ' ind2=' '>
-    <subfield code='a'>Graz, Techn. Univ., Diss.</subfield>
-    <subfield code='a'>2012</subfield>
-  </datafield>
-</record>
